refactor: split raw tokens
