+++ OUDS/Core/Tokens/RawTokens/Sources/BorderRawTokens.swift
@@ -37,39 +37,4 @@ public typealias BorderStyleRawToken = String
 ///  (i.e. publicly accessible from everywhere). More optimized than _struct_.
 ///
 /// In the future, generating tools like _Style Dictionary_ should either follow this file template and structure or just replace values or lines.
-public enum BorderRawTokens {
-
-    private static let borderBase: Double = 4 // Double type because used below for computations with Double values
-
-    // MARK: Primitive token - Border - Width
-
-    public static let borderWidth0: BorderWidthRawToken = borderBase * 0
-    public static let borderWidth25: BorderWidthRawToken = borderBase * 0.25
-    public static let borderWidth50: BorderWidthRawToken = borderBase * 0.5
-    public static let borderWidth75: BorderWidthRawToken = borderBase * 0.75
-    public static let borderWidth100: BorderWidthRawToken = borderBase * 1
-    public static let borderWidth150: BorderWidthRawToken = borderBase * 1.5
-    public static let borderWidth200: BorderWidthRawToken = borderBase * 2
-
-    // MARK: Primitive token - Border - Radius
-
-    public static let borderRadius0: BorderRadiusRawToken = borderBase * 0
-    public static let borderRadius25: BorderRadiusRawToken = borderBase * 0.25
-    public static let borderRadius50: BorderRadiusRawToken = borderBase * 0.5
-    public static let borderRadius75: BorderRadiusRawToken = borderBase * 0.75
-    public static let borderRadius100: BorderRadiusRawToken = borderBase * 1
-    public static let borderRadius150: BorderRadiusRawToken = borderBase * 1.5
-    public static let borderRadius200: BorderRadiusRawToken = borderBase * 2
-    public static let borderRadius300: BorderRadiusRawToken = borderBase * 3
-    public static let borderRadius400: BorderRadiusRawToken = borderBase * 4
-    public static let borderRadius500: BorderRadiusRawToken = borderBase * 5
-    public static let borderRadius600: BorderRadiusRawToken = borderBase * 6
-    public static let borderRadius800: BorderRadiusRawToken = borderBase * 8
-
-    // MARK: Primitive token - Border - Style
-
-    public static let borderStyleNone: BorderStyleRawToken = "none"
-    public static let borderStyleSolid: BorderStyleRawToken = "solid"
-    public static let borderStyleDashed: BorderStyleRawToken = "dashed"
-    public static let borderStyleDotted: BorderStyleRawToken = "dotted"
-}
+public enum BorderRawTokens { }
