+++ OUDS/Core/Tokens/RawTokens/Sources/OpacityRawTokens.swift
@@ -24,18 +24,4 @@ public typealias OpacityRawToken = Double
 /// Primitive types such as `Double` must be used to as to allow to use `@objc` keywords in extensions for overriding.
 /// Such tokens are packed in a _Swift enum_ so as to gather them in one object with the suitable namespace and avoid to have just constants in nothing else
 ///  (i.e. publicly accessible from everywhere). More optimized than _struct_.
-public enum OpacityRawTokens {
-
-    // MARK: Primitive token - Opacity
-
-    public static let opacity0: OpacityRawToken = 0
-    public static let opacity100: OpacityRawToken = 0.04
-    public static let opacity200: OpacityRawToken = 0.08
-    public static let opacity300: OpacityRawToken = 0.16
-    public static let opacity400: OpacityRawToken = 0.24
-    public static let opacity500: OpacityRawToken = 0.32
-    public static let opacity600: OpacityRawToken = 0.48
-    public static let opacity700: OpacityRawToken = 0.64
-    public static let opacity800: OpacityRawToken = 0.80
-    public static let opacity900: OpacityRawToken = 1
-}
+public enum OpacityRawTokens { }
