+++ OUDS/Core/Tokens/RawTokens/Sources/Values/DimenRawTokens-Values.swift
@@ -1,0 +1,56 @@
+//
+// Software Name: OUDS iOS
+// SPDX-FileCopyrightText: Copyright (c) Orange SA
+// SPDX-License-Identifier: MIT
+//
+// This software is distributed under the MIT license,
+// the text of which is available at https://opensource.org/license/MIT/
+// or see the "LICENSE" file for more details.
+//
+// Authors: See CONTRIBUTORS.txt
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
+//
+
+import Foundation
+
+/// Extracted in a separated files to help the *Figma* JSON to Swift parser to generate files to include easily.
+extension DimensionRawTokens {
+
+    public static let dimensionBase: DimensionRawToken = 4
+
+    // MARK: Primitive token - Dimension
+
+    public static let dimension0: DimensionRawToken = dimensionBase * 0
+    public static let dimension25: DimensionRawToken = dimensionBase * 0.5
+    public static let dimension50: DimensionRawToken = dimensionBase * 1
+    public static let dimension75: DimensionRawToken = dimensionBase * 1.5
+    public static let dimension100: DimensionRawToken = dimensionBase * 2
+    public static let dimension150: DimensionRawToken = dimensionBase * 3
+    public static let dimension200: DimensionRawToken = dimensionBase * 4
+    public static let dimension250: DimensionRawToken = dimensionBase * 5
+    public static let dimension300: DimensionRawToken = dimensionBase * 6
+    public static let dimension350: DimensionRawToken = dimensionBase * 7
+    public static let dimension400: DimensionRawToken = dimensionBase * 8
+    public static let dimension450: DimensionRawToken = dimensionBase * 9
+    public static let dimension500: DimensionRawToken = dimensionBase * 10
+    public static let dimension550: DimensionRawToken = dimensionBase * 11
+    public static let dimension600: DimensionRawToken = dimensionBase * 12
+    public static let dimension650: DimensionRawToken = dimensionBase * 13
+    public static let dimension700: DimensionRawToken = dimensionBase * 14
+    public static let dimension750: DimensionRawToken = dimensionBase * 15
+    public static let dimension800: DimensionRawToken = dimensionBase * 16
+    public static let dimension900: DimensionRawToken = dimensionBase * 18
+    public static let dimension1000: DimensionRawToken = dimensionBase * 20
+    public static let dimension1200: DimensionRawToken = dimensionBase * 24
+    public static let dimension1400: DimensionRawToken = dimensionBase * 28
+    public static let dimension1600: DimensionRawToken = dimensionBase * 32
+    public static let dimension1800: DimensionRawToken = dimensionBase * 36
+    public static let dimension2000: DimensionRawToken = dimensionBase * 40
+    public static let dimension3000: DimensionRawToken = dimensionBase * 80
+    public static let dimension4000: DimensionRawToken = dimensionBase * 120
+    public static let dimension5000: DimensionRawToken = dimensionBase * 140
+    public static let dimension6000: DimensionRawToken = dimensionBase * 160
+    public static let dimension7000: DimensionRawToken = dimensionBase * 180
+    public static let dimension9000: DimensionRawToken = dimensionBase * 220
+    public static let dimension11000: DimensionRawToken = dimensionBase * 260
+}
